Form-State-XML: Return success flag for XML-import on forms
The import of the state of a form via XML document is lenient and
ignores fields with errors. Such errors are only logged but some systems
need to know if a form state could be imported successfully or not. Add
a boolean result type which indicates whether the XML-import was
successful.

373731